backend/requirements-gemini-compatible.txt
# Gemini Environment Requirements (Python 3.13 Compatible)
# Fixed dependency conflicts

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.32.4

# Google AI (downgraded for compatibility)
google-generativeai==0.8.5
google-genai==1.0.0

# Audio processing
pydub==0.25.1
soundfile==0.13.1
librosa==0.11.0

# AI/ML
torch==2.7.1
numpy==2.2.6
pillow==11.3.0
tqdm==4.67.1

# Database
supabase==1.0.3

# Protocol buffers (compatible version)
protobuf==4.25.8 